Allow building the code with Clang 10 on Linux
Summary:
Allow building the code with Clang 10 on Linux.

Notes:
- We are using the "uninstrumented" directory instead of "clang_uninstrumented" for Clang-only third-party dependency archives. This is because when everything is built with Clang, we should not use the "clang_" prefix for just one directory in $YB_THIRDPARTY_DIR/installed.
- Adding YB_GCC_SUFFIX and YB_CLANG_SUFFIX variables to allow appending a suffix like "-<version>" to the compiler path.

Test Plan: Jenkins

Reviewers: jharveysmith, steve.varnau

Reviewed By: steve.varnau

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D9677

+++ src/yb/tablet/transaction_coordinator.cc
@@ -87,7 +87,8 @@ namespace tablet {
 std::chrono::microseconds GetTransactionTimeout() {
   const double timeout = GetAtomicFlag(&FLAGS_transaction_max_missed_heartbeat_periods) *
                          GetAtomicFlag(&FLAGS_transaction_heartbeat_usec);
-  return timeout >= std::chrono::microseconds::max().count()
+  // Cast to avoid -Wimplicit-int-float-conversion.
+  return timeout >= static_cast<double>(std::chrono::microseconds::max().count())
       ? std::chrono::microseconds::max()
       : std::chrono::microseconds(static_cast<int64_t>(timeout));
 }
